% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanframe.R
\name{cleanmerge}
\alias{cleanmerge}
\title{Cleaning the dataframe created by htbatch}
\usage{
cleanmerge(qsardataframe, ...)
}
\arguments{
\item{qsardataframe}{returned dataframe from htbatch}

\item{...}{inherit arguments}
}
\description{
Replaces dataframe NA values with zero and condenses htbatch returned dataframe to only relevant descriptor value columns for modeling.
}
\examples{
## Not run: cleanmerge(qsardataframe)
}
