% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataacidester.R
\docType{data}
\name{dataacidester}
\alias{dataacidester}
\title{Modeling descriptors for carboxylic acid esters}
\format{A dataframe with 70 chemical fragment sigma values and their rate constants formatted into 7 columns.}
\source{

}
\usage{
dataacidester
}
\description{
A dataset containing the fragment sigma values for 70 different carboxylic acid esters (specific carboxylic acid esters can be in found in acidesters.csv in inst/extdata) and the experimental rate constant.
}
\keyword{datasets}
