% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\docType{data}
\name{es}
\alias{es}
\title{Sigma values of steric fragments}
\format{A dataframe with 260 rows and 10 variables
\describe{
\item{sigma}{fragment SMILES, epi.value}
}}
\usage{
es
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation.
}
\keyword{datasets}
