% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{helper}
\title{Expedited search for phenyl fragment sigma values}
\usage{
helper(type, sigma.select)
}
\arguments{
\item{type}{Matches the HT.type specified in htbatch function value}

\item{sigma.select}{The type of sigma to be returned; valid inputs include "A", "B", "C", "D", "E", "F", "G", "H", and "U"}
}
\value{
values from library that are inserted directly into qsardataframe
}
\description{
This function is for returning values for the phenyl fragment without running htdesc.
}
