% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htdesc.R
\name{htdesc}
\alias{htdesc}
\title{HTdesc}
\usage{
htdesc(smile, HT.type = "taft", sigma.selection = "A", ...)
}
\arguments{
\item{smile}{SMILES string for a chemical fragment in character, factor, or SMIset datatype.}

\item{HT.type}{The type of Hammett-Taft (HT) descriptor; valid inputs include "taft", "meta", "para", "ortho", "inductive", "es" and "user" for a user defined sigma descriptor.}

\item{sigma.selection}{\itemize{The type of sigma to be returned; valid inputs include "A", "B", "C", "D", "E", "F", "G", "H", "Avg.dist", and "Med":
\item A = Avgerage value
\item B = Hansch preferred value if available then EPIsuite value if available then Average value
\item C = EPIsuite value if available then Hansch preferred value if available then Average of distinct values
\item D = EPIsuite value if available then Average of distinct value
\item E = Hansch preferred value if available then Average of distinct value
\item F = Hansch preferred value if available then mode value if available then median value
\item G = EPIsuite value if available then Hansch preferred value if available then median value
\item H = Mode if available then Average of distinct value
\item Avg.dist = Average of Distinct value
\item Med = Median Value
}}

\item{...}{arguments to be passed to fmcsR::fmcsBatch, such as al, au, bl, and bu}
}
\value{
A list containing the tanimoto coefficient for the closest matching MCS, SMILES string of the maximum common substructure (MCS), and index of the matched fragment from the library.
}
\description{
Get Hamett-Taft descriptor value for a single chemical fragment
}
\details{
Get values of Hammett-Taft descriptors for a given chemical fragment in SMILES string format by iterating through a lookup table.
Returns values from an exact matching fragment from the library or uses a substituted fragment.
In case an exact match isnt found, this function uses a mismatch-tolerant maximum common substructure (fMCS) based fragment substitution library to get the HT descriptors with highest tanimoto coefficient.
}
\examples{
htdesc (smile = "CCC", HT.type = "taft", sigma.selection = "A")

}
