% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htdescHelper.R
\docType{data}
\name{htdescHelper}
\alias{htdescHelper}
\title{Fast phenyl look up for htbatch}
\format{A dataframe with 24 phenyl sigma SMILES, their sigma values, location designation, and sigma.selection formatted into 7 columns.}
\source{

}
\usage{
htdescHelper
}
\description{
A dataset containing the phenyl sigma values for "taft", "steric", and "inductive" for all sigma.selection options.
}
\keyword{datasets}
