% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTdescR.R
\docType{package}
\name{HTdescR}
\alias{HTdescR}
\alias{HTdescR-package}
\title{HTdescR: A package for ...}
\description{
This package provides functionality for ...
}
\seealso{
Useful links:
\itemize{
  \item \url{http://jaympatel.com}
}

}
\author{
Jay M. Patel \email{jay@jaympatel.com}

Michaela Koopmans \email{michaela.koopmans@gmail.com}

John Olmstead \email{john.r.olmstead@gmail.com}
}
