% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indSDF.R
\docType{data}
\name{indSDF}
\alias{indSDF}
\title{FMCS searchable sigma values of inductive fragments}
\format{A Spatial Data File list of 274 s4 objects}
\source{

}
\usage{
indSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "indsigma" data frame.
}
\keyword{datasets}
