% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indsigma.R
\docType{data}
\name{indsigma}
\alias{indsigma}
\title{Sigma values of inductive fragments}
\format{A dataframe with 274 rows and 10 variables
\describe{
\item{sigma}{fragment SMILES, epi.value}
}}
\usage{
indsigma
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation.
}
\keyword{datasets}
