% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSDF.R
\docType{data}
\name{metaSDF}
\alias{metaSDF}
\title{FMCS searchable sigma values of meta substituted fragments}
\format{A Spatial Data File list of 373 s4 objects}
\source{

}
\usage{
metaSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "metasigma" data frame.
}
\keyword{datasets}
