% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoSDF.R
\docType{data}
\name{orthoSDF}
\alias{orthoSDF}
\title{FMCS searchable sigma values of ortho substituted fragments}
\format{A Spatial Data File list of 44 s4 objects}
\source{

}
\usage{
orthoSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "orthosigma" data frame.
}
\keyword{datasets}
