% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paraSDF.R
\docType{data}
\name{paraSDF}
\alias{paraSDF}
\title{FMCS searchable sigma values of para substituted fragments}
\format{A Spatial Data File list of 369 s4 objects}
\source{

}
\usage{
paraSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "parasigma" data frame.
}
\keyword{datasets}
