% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasigma.R
\docType{data}
\name{parasigma}
\alias{parasigma}
\title{Sigma values of para substituted fragments}
\format{A dataframe with 369 rows and 10 variables
\describe{
\item{sigma}{fragment SMILES, epi.value}
}}
\source{

}
\usage{
parasigma
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation.
}
\keyword{datasets}
