% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsarml.R
\name{plsqsar}
\alias{plsqsar}
\title{Predicting hydrolysis rates based on Quantitative Structure-Activity Relationships (QSAR) Models using Machine Learning Regressions (ML)}
\usage{
plsqsar(TrainingSet, TestSet, ...)
}
\arguments{
\item{...}{arguments from the fmcsR::fmcbatch such as a and b}

\item{smile}{SMILES string for a chemical fragment in character, factor, or SMIset datatype.}

\item{HT.type}{The type of Hammett-Taft (HT) descriptor; valid inputs include "taft", "meta", "para", "ortho", "induction", "es" and "user" for a user defined sigma descriptor.}
}
\value{
A list containing tanimoto coefficient for the closest matching MCS, SMILES string of the MCS, and index nuber of the matched fragment from the library.
}
\description{
Predict hydrolysis rates based on QSAR model using a variety of machine learning regressions such as random forest, support vector machine, partial least squares, and multiple linear regression.
}
\examples{
htdesc (smile = "CCC", HT.type = "taft", sigma.selection = "A")
}
