% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModels.R
\name{runModel}
\alias{runModel}
\title{Chemical prediction value generator}
\usage{
runModel(data, method)
}
\arguments{
\item{data}{chemical descriptor data frame}

\item{method}{The model from which the prediction should be generated.}
}
\value{
a data frame containing indeces, rate constant prediction, applicability domain info, and descriptors
}
\description{
R function that generates predictions for chemicals based on the machine learning method selected.
}
\examples{
##NOT RUN:runModel (dataacidester, method = "SVR")

}
