% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taftSDF.R
\docType{data}
\name{taftSDF}
\alias{taftSDF}
\title{FMCS searchable sigma values of taft fragments}
\format{A Spatial Data File list of 639 s4 objects}
\usage{
taftSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "taft" data frame.
}
\keyword{datasets}
