% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taftsigma.R
\docType{data}
\name{taftsigma}
\alias{taftsigma}
\title{Sigma values of taft fragments}
\format{A dataframe with 639 rows and 10 variables
\describe{
 \item{sigma}{fragment SMILES, epi.value}
 ...
}}
\source{

}
\usage{
taftsigma
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation.
}
\keyword{datasets}
