% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userSDF.R
\docType{data}
\name{userSDF}
\alias{userSDF}
\title{FMCS searchable sigma values of steric fragments}
\format{A Spatial Data File list of 260 s4 objects}
\source{

}
\usage{
userSDF
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation. fmcs searchable component to the "es" data frame.
}
\keyword{datasets}
