% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersigma.R
\docType{data}
\name{usersigma}
\alias{usersigma}
\title{Sigma values of ortho substituted fragments}
\format{A dataframe with 44 rows and 10 variables
\describe{
\item{sigma}{fragment SMILES, epi.value}
}}
\usage{
usersigma
}
\description{
A dataset containing the fragment SMILES and various sigma values coming from datasets or
calculated from the dataset amalgamation.
}
\keyword{datasets}
