% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htdesc.R
\name{htdesc}
\alias{htdesc}
\title{Get Hamett-Taft descriptor for a single chemical fragment}
\usage{
htdesc(smile, HT.type = "taft", sigma.selection = "A", ...)
}
\arguments{
\item{smile}{SMILES string for a chemical fragment in character, factor, or SMIset datatype.}

\item{HT.type}{The type of Hammett-Taft (HT) descriptor; valid inputs include "taft", "meta", "para", "ortho", "induction", "es" and "user" for a user defined sigma descriptor.}

\item{sigma.selection}{The type of sigma to be returned; valid inputs include "A", "B", "C", "D", "E", "F", "G", "H", and "U"}

\item{...}{arguments from the fmcsR::fmcbatch such as a and b}
}
\value{
A list containing tanimoto coefficient for the closest matching MCS, SMILES string of the MCS, and index nuber of the matched fragment from the library.
}
\description{
Get values of Hammett-Taft descriptors for a given chemical fragment in SMILES string format by iterating through a lookup table. In case an exact match isnt found, this function uses a mismatch tolerant
maximum common substructure (fMCS) based fragment substitution library to get the HT dexcriptors with highest tanimoto coefficient.
}
\examples{
htdesc (smile = "CCC", HT.type = "taft", sigma.selection = "A")

}
